requirements.txt
Django==1.8
django-filter==0.10.0
djangorestframework==3.1.2
